chore: [V2] rename 'core' module to 'common' (#1364)
* First rename

* POM updates

* Rename namespace

* Fix

* Fix?

* More namespace changes

* Rename core examples to core utilities

* Finish renaming core utilities

* Clean up name/description of module

* More change

* Fix readme
