+++ examples/powertools-examples-core-utilities/cdk/README.md
@@ -6,11 +6,11 @@ For general information on the deployed example itself, you can refer to the par
 
 ## Configuration
 CDK uses the following project structure:
-- [app](./app) - stores the source code of your application, which is similar between all examples
-- [infra](./infra) - stores the definition of your infrastructure
-  - [cdk.json](./infra/cdk.json) - tells the CDK Toolkit how to execute your app
-  - [CdkApp](./infra/src/main/java/cdk/CdkApp.java) - bootstraps your stack, taking AWS `account` and `region` as input
-  - [CdkStack](./infra/src/main/java/cdk/CdkStack.java) - defines the Lambda function to be deployed as well as API Gateway for it.
+- [app](app) - stores the source code of your application, which is similar between all examples
+- [infra](infra) - stores the definition of your infrastructure
+  - [cdk.json](infra/cdk.json) - tells the CDK Toolkit how to execute your app
+  - [CdkApp](infra/src/main/java/cdk/CdkApp.java) - bootstraps your stack, taking AWS `account` and `region` as input
+  - [CdkStack](infra/src/main/java/cdk/CdkStack.java) - defines the Lambda function to be deployed as well as API Gateway for it.
 
 It is a [Maven](https://maven.apache.org/) based project, so you can open this project with any Maven compatible Java IDE to build and run tests.
