+++ examples/powertools-examples-core-utilities/cdk/app/events/event.json
