+++ examples/powertools-examples-core-utilities/cdk/app/pom.xml
