+++ examples/powertools-examples-core-utilities/cdk/app/src/main/java/helloworld/AppStream.java
