+++ examples/powertools-examples-core-utilities/cdk/app/src/main/resources/log4j2.xml
