+++ examples/powertools-examples-core-utilities/cdk/infra/cdk.json
