+++ examples/powertools-examples-core-utilities/cdk/infra/pom.xml
