+++ examples/powertools-examples-core-utilities/cdk/infra/src/main/java/cdk/CdkStack.java
