+++ examples/powertools-examples-core-utilities/cdk/infra/src/test/java/cdk/CdkStackTest.java
