+++ examples/powertools-examples-core-utilities/sam/events/event.json
