+++ examples/powertools-examples-core-utilities/sam/pom.xml
@@ -6,8 +6,8 @@
     <version>2.0.0-SNAPSHOT</version>
     <artifactId>powertools-examples-core-sam</artifactId>
     <packaging>jar</packaging>
-
-    <name>Powertools for AWS Lambda (Java) library Examples - Core</name>
+    
+    <name>Powertools for AWS Lambda (Java) library Examples - Core Utilities (logging, tracing, metrics)</name>
 
     <properties>
         <log4j.version>2.20.0</log4j.version>
