+++ examples/powertools-examples-core-utilities/sam/src/main/java/helloworld/AppStream.java
