+++ examples/powertools-examples-core-utilities/sam/src/main/resources/log4j2.xml
