+++ examples/powertools-examples-core-utilities/sam/template.yaml
