+++ powertools-common/src/main/resources-filtered/version.properties
