+++ powertools-common/src/test/java/software/amazon/lambda/powertools/common/internal/UserAgentConfiguratorTest.java
@@ -12,15 +12,16 @@
  *
  */
 
-package software.amazon.lambda.powertools.core.internal;
+package software.amazon.lambda.powertools.common.internal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mockStatic;
-import static software.amazon.lambda.powertools.core.internal.SystemWrapper.getenv;
-import static software.amazon.lambda.powertools.core.internal.UserAgentConfigurator.AWS_EXECUTION_ENV;
-import static software.amazon.lambda.powertools.core.internal.UserAgentConfigurator.VERSION_KEY;
-import static software.amazon.lambda.powertools.core.internal.UserAgentConfigurator.VERSION_PROPERTIES_FILENAME;
-import static software.amazon.lambda.powertools.core.internal.UserAgentConfigurator.getVersionFromProperties;
+import static software.amazon.lambda.powertools.common.internal.UserAgentConfigurator.AWS_EXECUTION_ENV;
+import static software.amazon.lambda.powertools.common.internal.UserAgentConfigurator.VERSION_KEY;
+import static software.amazon.lambda.powertools.common.internal.UserAgentConfigurator.VERSION_PROPERTIES_FILENAME;
+import static software.amazon.lambda.powertools.common.internal.UserAgentConfigurator.getVersionFromProperties;
+import static software.amazon.lambda.powertools.common.internal.SystemWrapper.getenv;
+
 
 import java.io.File;
 import java.util.Objects;
