+++ powertools-common/src/test/resources/test.properties
