+++ powertools-common/src/test/resources/unreadable.properties
