+++ powertools-large-messages/pom.xml
@@ -56,7 +56,7 @@
     <dependencies>
         <dependency>
             <groupId>software.amazon.lambda</groupId>
-            <artifactId>powertools-core</artifactId>
+            <artifactId>powertools-common</artifactId>
         </dependency>
         <dependency>
             <groupId>org.aspectj</groupId>
